Added similar_words + finalised multiplechoice
