+++ src/api/exercises.js
@@ -16,6 +16,6 @@ Zeeguu_API.prototype.uploadExerciseFeedback = function (
   );
 };
 
-Zeeguu_API.prototype.wordsSimilarTo = function (bookmark_id) {
-  return ["foo", "bar"];
+Zeeguu_API.prototype.wordsSimilarTo = function (bookmark_id, callback) {
+  this._get(`similar_words/${bookmark_id}`, callback);
 };
