+++ src/exercises/exerciseTypes/findWordInContext/BottomInput.js
